Template Ant Build
The WebWorks 2.0 Cordova plugin template requires developers to
manually update the template code to remove references to template.
This is a very manual process and is error prone. The ant build allows
the developer to create a fresh project to start coding without the
hassle of removing all of the template references.