Added Questionnaire support
Added questionnaire prompts to the build script. The user no longer
needs to provide input parameters. Since importing the project into
Momentics and configuring not to place the project contents into the
workspace the build will now exit when it determines there is an
existing build directory and the user selects not to clean the build
directory. This is to safe guard any deletion of project contents.