Moved config.xml outside of www folder
Moved the config.xml out side of the www folder because webworks now
creates the config.xml outside of the www folder. The sample config.xml
widget id is now com.sample.PROJECT_NAME_LOWERCASE in case a user
creates the plugin as com.blackberry.community.xyz blackberry is a
reserved word in the widget id.